refactor(test): move MSW server setup to global test-setup.ts
Resolves the TODO left by @richieforeman in config.integration.test.ts.
The MSW server is now initialised once in the global test-setup file so
it is available to all future CLI package tests without duplication.

- Add `setupServer()`, `beforeAll/afterAll` lifecycle hooks and
  `server.resetHandlers()` in the shared `afterEach` to `test-setup.ts`.
- Export `server` from `test-setup.ts` so individual test files can
  register their own handlers via `server.resetHandlers(...)`.
- Remove the redundant MSW bootstrap code from
  `config.integration.test.ts` and import `server` from the global
  setup file instead.

Fixes #21660
